requirements.txt
django
psycopg2-binary
drf-spectacular
djangorestframework
pytest
pytest-django